refactor(react-ui-stack): reorganize internals, audit consumers, drop stale deps
- Reorganize src/components: Stack/, StackContext/, StackItem/, StackItemSigil/
  modules; consolidate Orientation/Size/StackContextValue into types.ts; move
  hooks into Stack/. Public export surface unchanged.
- Add AUDIT.md documenting the public API surface, internal structure, and the
  remaining consumers (plugin-deck, plugin-stack, plugin-script, plugin-navtree).
- Drop the @dxos/react-ui-stack dependency + tsconfig project ref from 10 packages
  that no longer use it: plugin-explorer, plugin-registry, react-ui-canvas-compute,
  react-ui-graph, react-ui-masonry, react-ui-table (stale), stories-assistant
  (stories removed), plugin-board, plugin-pipeline, react-ui-board (translations
  merge removed).
- fix(plugin-script): guard Obj.getDatabase against an undefined notebook in
  NotebookArticle (crashed the NotebookArticle story on first render).
