+++ packages/plugins/plugin-board/tsconfig.json
@@ -70,9 +70,6 @@
       "path": "../../ui/react-ui-form"
     },
     {
-      "path": "../../ui/react-ui-stack"
-    },
-    {
       "path": "../../ui/react-ui-syntax-highlighter"
     },
     {
