+++ packages/plugins/plugin-explorer/package.json
@@ -111,7 +111,6 @@
     "@dxos/react-ui-components": "workspace:*",
     "@dxos/react-ui-graph": "workspace:*",
     "@dxos/react-ui-mosaic": "workspace:*",
-    "@dxos/react-ui-stack": "workspace:*",
     "@dxos/schema": "workspace:*",
     "@dxos/types": "workspace:*",
     "@dxos/util": "workspace:*",
