+++ packages/plugins/plugin-explorer/tsconfig.json
@@ -84,9 +84,6 @@
       "path": "../../ui/react-ui-mosaic"
     },
     {
-      "path": "../../ui/react-ui-stack"
-    },
-    {
       "path": "../../ui/ui-theme"
     },
     {
