+++ packages/plugins/plugin-pipeline/tsconfig.json
@@ -88,9 +88,6 @@
       "path": "../../ui/react-ui-mosaic"
     },
     {
-      "path": "../../ui/react-ui-stack"
-    },
-    {
       "path": "../../ui/ui-theme"
     },
     {
