+++ packages/plugins/plugin-registry/tsconfig.json
@@ -85,9 +85,6 @@
       "path": "../../ui/react-ui-mosaic"
     },
     {
-      "path": "../../ui/react-ui-stack"
-    },
-    {
       "path": "../../ui/ui-theme"
     },
     {
