--- packages/stories/stories-assistant/src/stories/Projects.stories.tsx
@@ -1,166 +1,0 @@
-//
-// Copyright 2025 DXOS.org
-//
-
-import { type Meta, type StoryObj } from '@storybook/react-vite';
-import * as Effect from 'effect/Effect';
-import React, { type FC, useCallback } from 'react';
-
-import { Capabilities } from '@dxos/app-framework';
-import { Surface, useCapabilities, useCapability } from '@dxos/app-framework/ui';
-import { AppCapabilities } from '@dxos/app-toolkit';
-import { AppSurface } from '@dxos/app-toolkit/ui';
-import { AiContext } from '@dxos/assistant';
-import { Blueprint } from '@dxos/compute';
-import { Database, Filter, Obj, Ref } from '@dxos/echo';
-import { makeRegistry } from '@dxos/echo-client';
-import { EffectEx } from '@dxos/effect';
-import { log } from '@dxos/log';
-import { Assistant } from '@dxos/plugin-assistant';
-import { useContextBinder } from '@dxos/plugin-assistant/hooks';
-import { translations } from '@dxos/plugin-assistant/translations';
-import { MarkdownPlugin } from '@dxos/plugin-markdown/testing';
-import { useQuery, useSpaces } from '@dxos/react-client/echo';
-import { useAsyncEffect } from '@dxos/react-ui';
-import { Stack, StackItem } from '@dxos/react-ui-stack';
-import { withLayout, withTheme, Loading } from '@dxos/react-ui/testing';
-import { isNonNullable } from '@dxos/util';
-
-import { ChatModule, type ModuleProps } from '../components';
-import { config, getDecorators } from '../testing';
-
-// TODO(burdon): Move into Chat.stories.tsx
-
-const panelClassNames = 'bg-base-surface rounded-xs border border-separator overflow-hidden';
-
-type DefaultStoryProps = {
-  modules: FC<ModuleProps>[][];
-  blueprints?: string[];
-  showContext?: boolean;
-};
-
-const DefaultStory = ({ modules, showContext, blueprints = [] }: DefaultStoryProps) => {
-  const blueprintsDefinitions = useCapabilities(AppCapabilities.BlueprintDefinition);
-  const atomRegistry = useCapability(Capabilities.AtomRegistry);
-
-  const [space] = useSpaces();
-  useAsyncEffect(async () => {
-    if (!space) {
-      return;
-    }
-
-    const chats = await space.db.query(Filter.type(Assistant.Chat)).run();
-    const chat = chats[0];
-    if (!chat) {
-      return;
-    }
-
-    // Add blueprints to context.
-    const registry = makeRegistry({ initial: blueprintsDefinitions.map((def) => def.make()) });
-    const blueprintObjects = blueprints
-      .map((key) => {
-        const blueprint = registry
-          .query(Filter.type(Blueprint.Blueprint))
-          .runSync()
-          .find((b) => Obj.getMeta(b).key === key);
-        if (blueprint) {
-          return space.db.add(Obj.clone(blueprint));
-        }
-      })
-      .filter(isNonNullable);
-
-    const feedTarget = await chat.feed.load();
-    const runtime = await EffectEx.runAndForwardErrors(
-      Effect.runtime<Database.Service>().pipe(Effect.provide(Database.layer(space.db))),
-    );
-    const binder = new AiContext.Binder({ feed: feedTarget, runtime, registry: atomRegistry });
-    await binder.use((binder) => binder.bind({ blueprints: blueprintObjects.map((blueprint) => Ref.make(blueprint)) }));
-  }, [space, blueprints, blueprintsDefinitions]);
-
-  const handleEvent = useCallback<NonNullable<ModuleProps['onEvent']>>((event) => {
-    log.info('event', { event });
-  }, []);
-
-  const chats = useQuery(space?.db, Filter.type(Assistant.Chat));
-  const feedTarget = chats.at(-1)?.feed.target;
-  const binder = useContextBinder(space, feedTarget);
-  const objects = binder?.getObjects() ?? [];
-
-  if (!space) {
-    return <Loading data={{ space: !!space }} />;
-  }
-
-  return (
-    <Stack
-      orientation='horizontal'
-      size='split'
-      rail={false}
-      itemsCount={modules.length + (showContext ? 1 : 0)}
-      classNames='absolute inset-0 gap-(--stack-gap)'
-    >
-      {modules.map((Components, i) => {
-        return (
-          <StackItem.Root key={i} item={{ id: `${i}` }}>
-            <Stack
-              orientation='vertical'
-              classNames='gap-(--stack-gap)'
-              size={i > 0 ? 'contain' : 'split'}
-              itemsCount={Components.length}
-              rail={false}
-            >
-              {Components.map((Component, i) => (
-                <StackItem.Root key={i} item={{ id: `${i}` }} classNames={panelClassNames}>
-                  <Component space={space} onEvent={handleEvent} />
-                </StackItem.Root>
-              ))}
-            </Stack>
-          </StackItem.Root>
-        );
-      })}
-
-      {showContext && <StackContainer objects={objects} />}
-    </Stack>
-  );
-};
-
-const StackContainer = ({ objects }: { objects: Obj.Any[] }) => {
-  return (
-    <Stack
-      orientation='vertical'
-      classNames='gap-(--stack-gap)'
-      size='contain'
-      rail={false}
-      itemsCount={objects.length}
-    >
-      {objects.map((object) => (
-        <StackItem.Root key={object.id} item={object} classNames={panelClassNames}>
-          <Surface.Surface type={AppSurface.Section} limit={1} data={{ subject: object, attendableId: object.id }} />
-        </StackItem.Root>
-      ))}
-    </Stack>
-  );
-};
-
-const storybook: Meta<typeof DefaultStory> = {
-  title: 'stories/stories-assistant/Projects',
-  render: DefaultStory,
-  decorators: [withTheme(), withLayout({ layout: 'fullscreen' })],
-  parameters: {
-    layout: 'fullscreen',
-    translations,
-  },
-};
-
-export default storybook;
-
-type Story = StoryObj<typeof storybook>;
-
-export const Default: Story = {
-  decorators: getDecorators({
-    plugins: [MarkdownPlugin()],
-    config: config.remote,
-  }),
-  args: {
-    modules: [[ChatModule]],
-  },
-};
