--- packages/stories/stories-assistant/src/stories/QueryEditor.stories.tsx
@@ -1,104 +1,0 @@
-//
-// Copyright 2025 DXOS.org
-//
-
-import { type Meta, type StoryObj } from '@storybook/react-vite';
-import React, { useState } from 'react';
-
-import { Obj, Ref, Tag } from '@dxos/echo';
-import { translations } from '@dxos/plugin-assistant/translations';
-import { ForceGraph } from '@dxos/plugin-explorer/components';
-import { useGraphModel } from '@dxos/plugin-explorer/hooks';
-import { random } from '@dxos/random';
-import { useQuery } from '@dxos/react-client/echo';
-import { useClientStory, withClientProvider } from '@dxos/react-client/testing';
-import { ScrollArea } from '@dxos/react-ui';
-import { QueryEditor, type QueryEditorProps, useQueryBuilder } from '@dxos/react-ui-components';
-import { withLayout, withTheme } from '@dxos/react-ui/testing';
-import { type ValueGenerator, createObjectFactory } from '@dxos/schema/testing';
-import { Employer, Organization, Person, Pipeline } from '@dxos/types';
-
-// TODO(burdon): Move.
-
-random.seed(1);
-const generator = random as any as ValueGenerator;
-
-const DefaultStory = ({ value: valueProp }: QueryEditorProps) => {
-  const { space } = useClientStory();
-  const [query, setQuery] = useState<string | undefined>(valueProp);
-  const filter = useQueryBuilder(query);
-  const objects = useQuery(space?.db, filter).sort(Obj.sort(Obj.sortByTypename, Obj.sortByLabel));
-  const model = useGraphModel(space?.db, filter);
-
-  return (
-    <div className='grid grid-cols-2 grow divide-x divide-subdued-separator overflow-hidden'>
-      <div className='flex flex-col overflow-hidden'>
-        <QueryEditor classNames='p-2 w-full border-b border-subdued-separator' db={space?.db} onChange={setQuery} />
-        <ScrollArea.Root orientation='vertical'>
-          <ScrollArea.Viewport>
-            {objects.map((object) => (
-              <div
-                key={object.id}
-                className='grid grid-cols-3 gap-2 p-2 border-b border-subdued-separator overflow-hidden'
-              >
-                <span className='truncate text-sm font-mono'>{object.id}</span>
-                <span className='truncate text-sm font-mono'>{Obj.getTypename(object)}</span>
-                <span className='truncate'>{Obj.getLabel(object)}</span>
-              </div>
-            ))}
-          </ScrollArea.Viewport>
-        </ScrollArea.Root>
-        <div className='p-2 text-right text-info-text text-xs'>{objects.length}</div>
-      </div>
-      <ForceGraph model={model} />
-    </div>
-  );
-};
-
-const meta: Meta<typeof QueryEditor> = {
-  title: 'stories/stories-assistant/QueryEditor',
-  component: QueryEditor,
-  render: DefaultStory,
-  decorators: [
-    withTheme(),
-    withLayout({ layout: 'fullscreen' }),
-    withClientProvider({
-      types: [Organization.Organization, Person.Person, Pipeline.Pipeline, Employer.Employer, Tag.Tag],
-      createIdentity: true,
-      createSpace: true,
-      onCreateSpace: async ({ space }) => {
-        // Persist real Tag objects so the seeded `meta.tags` refs resolve like production data.
-        const tags = [
-          space.db.add(Tag.make({ label: 'Red' })),
-          space.db.add(Tag.make({ label: 'Green' })),
-          space.db.add(Tag.make({ label: 'Blue' })),
-        ];
-        const createObjects = createObjectFactory(space.db, generator);
-        const objects = await createObjects([
-          { type: Organization.Organization, count: 30 },
-          { type: Pipeline.Pipeline, count: 20 },
-          { type: Person.Person, count: 50 },
-        ]);
-        objects.forEach((obj) => {
-          Obj.update(obj, (obj) => {
-            Obj.getMeta(obj).tags = random.helpers.uniqueArray(tags, random.number.int(3)).map((tag) => Ref.make(tag));
-          });
-        });
-      },
-    }),
-  ],
-  parameters: {
-    layout: 'fullscreen',
-    translations,
-  },
-};
-
-export default meta;
-
-type Story = StoryObj<typeof meta>;
-
-export const Default: Story = {
-  args: {
-    value: '(type:org.dxos.type.person OR type:org.dxos.type.organization)',
-  },
-};
