+++ packages/ui/react-ui-board/tsconfig.json
@@ -54,9 +54,6 @@
       "path": "../react-ui-mosaic"
     },
     {
-      "path": "../react-ui-stack"
-    },
-    {
       "path": "../react-ui-syntax-highlighter"
     },
     {
