+++ packages/ui/react-ui-canvas-compute/tsconfig.json
@@ -120,9 +120,6 @@
       "path": "../react-ui-mosaic"
     },
     {
-      "path": "../react-ui-stack"
-    },
-    {
       "path": "../react-ui-syntax-highlighter"
     },
     {
