+++ packages/ui/react-ui-graph/tsconfig.json
@@ -39,9 +39,6 @@
       "path": "../react-ui-mosaic"
     },
     {
-      "path": "../react-ui-stack"
-    },
-    {
       "path": "../react-ui-syntax-highlighter"
     },
     {
