+++ packages/ui/react-ui-masonry/tsconfig.json
@@ -48,9 +48,6 @@
       "path": "../react-ui-mosaic"
     },
     {
-      "path": "../react-ui-stack"
-    },
-    {
       "path": "../ui-theme"
     },
     {
