+++ packages/ui/react-ui-stack/src/components/StackContext/index.ts
@@ -1,0 +1,5 @@
+//
+// Copyright 2024 DXOS.org
+//
+
+export * from './StackContext';
