+++ packages/ui/react-ui-stack/src/components/StackItemSigil/StackItemSigil.tsx
