+++ packages/ui/react-ui-stack/src/components/index.ts
@@ -5,5 +5,5 @@
 export type * from './types';
 
 export * from './Stack';
-
+export * from './StackContext';
 export * from './StackItem';
