--- packages/ui/react-ui-stack/src/hooks/index.ts
@@ -1,5 +1,0 @@
-//
-// Copyright 2025 DXOS.org
-//
-
-export { useStackDropForElements } from './useStackDropForElements';
