+++ packages/ui/react-ui-table/tsconfig.json
@@ -90,9 +90,6 @@
       "path": "../react-ui-search"
     },
     {
-      "path": "../react-ui-stack"
-    },
-    {
       "path": "../react-ui-syntax-highlighter"
     },
     {
